Analysis of RTU runtime for 1- and 2-stage cooling
Comparison of the RTU runtime for zones selected buildings in the XBOS project. The buildings and zones included here are those that had RTUs with 2-stage cooling AND were received site permission for experimentation. The analysis shows the runtime for each state in hours for each of the sites, zones, and PDP or CPP event dates.